+++ serialization/src/commonMain/kotlin/nl/adaptivity/xmlutil/serialization/XMLDecoder.kt
@@ -128,9 +128,12 @@ internal open class XmlDecoderBase internal constructor(
 
     abstract inner class DecodeCommons(
         xmlDescriptor: XmlDescriptor,
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
     ) : XmlCodec<XmlDescriptor>(xmlDescriptor), XML.XmlInput, Decoder {
         final override val config: XmlConfig get() = this@XmlDecoderBase.config
         final override val serializersModule: SerializersModule get() = this@XmlDecoderBase.serializersModule
+
+        protected val preserveSpace: DocumentPreserveSpace = inheritedPreserveWhitespace.withDefault(xmlDescriptor.defaultPreserveSpace)
 
         override fun decodeNull(): Nothing? {
             // We don't write nulls, so if we know that we have a null we just return it
@@ -215,7 +218,8 @@ internal open class XmlDecoderBase internal constructor(
         protected val polyInfo: PolyInfo? = null,
         protected val isValueChild: Boolean = false,
         val attrIndex: Int = -1,
-    ) : DecodeCommons(xmlDescriptor), Decoder, XML.XmlInput, ChunkedDecoder {
+        inheritedPreserveWhitespace: DocumentPreserveSpace
+    ) : DecodeCommons(xmlDescriptor, inheritedPreserveWhitespace), Decoder, XML.XmlInput, ChunkedDecoder {
         final override val input: XmlPeekingReader get() = this@XmlDecoderBase.input
 
         private var triggerInline = false
@@ -264,7 +268,7 @@ internal open class XmlDecoderBase internal constructor(
 
                     OutputKind.Inline -> throw SerializationException("Inline classes can not be decoded directly")
                     OutputKind.Mixed -> input.allConsecutiveTextContent()//.also { input.next() } // Move to the next element
-                    OutputKind.Text -> if (xmlDescriptor.preserveSpace) {
+                    OutputKind.Text -> if (xmlDescriptor.defaultPreserveSpace.withDefault(true)) {
                         input.allConsecutiveTextContent()
                     } else {
                         input.allText()
@@ -303,7 +307,7 @@ internal open class XmlDecoderBase internal constructor(
 
                     OutputKind.Inline -> throw SerializationException("Inline classes can not be decoded directly")
                     OutputKind.Mixed -> return input.allConsecutiveTextContentChunked(consumeChunk)
-                    OutputKind.Text -> return if (xmlDescriptor.preserveSpace) {
+                    OutputKind.Text -> return if (xmlDescriptor.defaultPreserveSpace.withDefault(true)) {
                         input.allConsecutiveTextContentChunked(consumeChunk)
                     } else {
                         input.allTextChunked(consumeChunk)
@@ -338,7 +342,7 @@ internal open class XmlDecoderBase internal constructor(
             val startsWithTag = input.eventType == EventType.START_ELEMENT
             val startDepth = input.depth
             val serialValueDecoder =
-                SerialValueDecoder(effectiveDeserializer, desc, polyInfo, attrIndex, typeDiscriminatorName, isValueChild)
+                SerialValueDecoder(effectiveDeserializer, desc, polyInfo, attrIndex, typeDiscriminatorName, isValueChild, preserveSpace)
             val value = effectiveDeserializer.deserializeSafe(serialValueDecoder)
             if (startsWithTag && !input.hasPeekItems && input.depth < startDepth) {
                 input.pushBackCurrent()
@@ -357,8 +361,9 @@ internal open class XmlDecoderBase internal constructor(
     internal inner class StringDecoder(
         xmlDescriptor: XmlDescriptor,
         private val locationInfo: XmlReader.LocationInfo?,
-        private val stringValue: String
-    ) : Decoder, XML.XmlInput, DecodeCommons(xmlDescriptor) {
+        private val stringValue: String,
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : Decoder, XML.XmlInput, DecodeCommons(xmlDescriptor, inheritedPreserveWhitespace) {
         override val input: XmlPeekingReader by lazy { PseudoBufferedReader(XmlStringReader(locationInfo, stringValue)) }
 
         override fun beginStructure(descriptor: SerialDescriptor): CompositeDecoder {
@@ -370,7 +375,7 @@ internal open class XmlDecoderBase internal constructor(
 
         @ExperimentalSerializationApi
         override fun decodeInline(descriptor: SerialDescriptor): Decoder {
-            return StringDecoder(xmlDescriptor.getElementDescriptor(0), locationInfo, stringValue)
+            return StringDecoder(xmlDescriptor.getElementDescriptor(0), locationInfo, stringValue, preserveSpace)
         }
 
         override fun decodeStringImpl(defaultOverEmpty: Boolean): String {
@@ -494,8 +499,9 @@ internal open class XmlDecoderBase internal constructor(
         polyInfo: PolyInfo?,
         attrIndex: Int,/* = -1*/
         override val typeDiscriminatorName: QName?,
-        isValueChild: Boolean
-    ) : XmlDecoder(xmlDescriptor, polyInfo, isValueChild, attrIndex) {
+        isValueChild: Boolean,
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : XmlDecoder(xmlDescriptor, polyInfo, isValueChild, attrIndex, inheritedPreserveWhitespace) {
         private var notNullChecked = false
 
         /** Object that allows recording id attributes */
@@ -539,7 +545,8 @@ internal open class XmlDecoderBase internal constructor(
             if (descriptor.isNullable) return TagDecoder(
                 deserializer,
                 xmlDescriptor,
-                typeDiscriminatorName
+                typeDiscriminatorName,
+                preserveSpace
             ).also { tagIdHolder = it }
 
             return when {
@@ -547,7 +554,7 @@ internal open class XmlDecoderBase internal constructor(
                     throw AssertionError("A primitive is not a composite")
 
                 xmlDescriptor is XmlPolymorphicDescriptor ->
-                    PolymorphicDecoder(deserializer, xmlDescriptor, polyInfo, isValueChild).also { tagIdHolder = it }
+                    PolymorphicDecoder(deserializer, xmlDescriptor, polyInfo, isValueChild, preserveSpace).also { tagIdHolder = it }
 
                 xmlDescriptor is XmlListDescriptor -> when {
                     xmlDescriptor.outputKind == OutputKind.Attribute ->
@@ -555,21 +562,34 @@ internal open class XmlDecoderBase internal constructor(
                             deserializer,
                             xmlDescriptor,
                             input.extLocationInfo,
-                            attrIndex
+                            attrIndex,
+                            preserveSpace,
                         ).also { tagIdHolder = it }
 
-                    xmlDescriptor.outputKind == OutputKind.Text ->
-                        ValueListDecoder(deserializer, xmlDescriptor, input.extLocationInfo)
+                    xmlDescriptor.outputKind == OutputKind.Text -> ValueListDecoder(
+                        deserializer,
+                        xmlDescriptor,
+                        input.extLocationInfo,
+                        preserveSpace
+                    )
 
                     xmlDescriptor.isListEluded ->
-                        AnonymousListDecoder(deserializer, xmlDescriptor, polyInfo, typeDiscriminatorName, isValueChild).also {
+                        AnonymousListDecoder(
+                            deserializer,
+                            xmlDescriptor,
+                            polyInfo,
+                            typeDiscriminatorName,
+                            isValueChild,
+                            preserveSpace
+                        ).also {
                             tagIdHolder = it
                         }
 
                     else -> NamedListDecoder(
                         deserializer,
                         xmlDescriptor,
-                        typeDiscriminatorName
+                        typeDiscriminatorName,
+                        preserveSpace
                     ).also { tagIdHolder = it }
                 }
 
@@ -579,19 +599,21 @@ internal open class XmlDecoderBase internal constructor(
                             deserializer,
                             xmlDescriptor,
                             polyInfo,
-                            typeDiscriminatorName
+                            typeDiscriminatorName,
+                            preserveSpace,
                         ).also { tagIdHolder = it }
 
                     else -> NamedMapDecoder(
                         deserializer,
                         xmlDescriptor,
                         polyInfo,
-                        typeDiscriminatorName
+                        typeDiscriminatorName,
+                        preserveSpace,
                     ).also { tagIdHolder = it }
 
                 }
 
-                else -> TagDecoder(deserializer, xmlDescriptor, typeDiscriminatorName).also { tagIdHolder = it }
+                else -> TagDecoder(deserializer, xmlDescriptor, typeDiscriminatorName, preserveSpace).also { tagIdHolder = it }
             }.also {
                 for (attrName in ignoredAttributes) {
                     it.ignoreAttribute(attrName)
@@ -606,7 +628,7 @@ internal open class XmlDecoderBase internal constructor(
      * are actually parsed as XML and can be complex
      */
     private inner class NullDecoder(xmlDescriptor: XmlDescriptor, isValueChild: Boolean) :
-        XmlDecoder(xmlDescriptor, isValueChild = isValueChild), CompositeDecoder {
+        XmlDecoder(xmlDescriptor, isValueChild = isValueChild, inheritedPreserveWhitespace = DocumentPreserveSpace.DEFAULT), CompositeDecoder {
 
         override fun decodeNotNullMark() = (xmlDescriptor as? XmlValueDescriptor)?.default != null
 
@@ -700,7 +722,8 @@ internal open class XmlDecoderBase internal constructor(
         deserializer: DeserializationStrategy<*>,
         xmlDescriptor: D,
         typeDiscriminatorName: QName?,
-    ) : TagDecoderBase<D>(deserializer, xmlDescriptor, typeDiscriminatorName) {
+        preserveWhitespace: DocumentPreserveSpace,
+    ) : TagDecoderBase<D>(deserializer, xmlDescriptor, typeDiscriminatorName, preserveWhitespace) {
 
         private val readTagName = if(config.isUnchecked) xmlDescriptor.tagName else input.name
 
@@ -726,7 +749,8 @@ internal open class XmlDecoderBase internal constructor(
     internal abstract inner class TagDecoderBase<D : XmlDescriptor>(
         protected val deserializer: DeserializationStrategy<*>,
         xmlDescriptor: D,
-        protected val typeDiscriminatorName: QName?
+        protected val typeDiscriminatorName: QName?,
+        preserveWhitespace: DocumentPreserveSpace,
     ) : XmlTagCodec<D>(xmlDescriptor), CompositeDecoder, XML.XmlInput, TagIdHolder {
 
         override var tagId: String? = null
@@ -736,7 +760,8 @@ internal open class XmlDecoderBase internal constructor(
          * Determine whether whitespace is preserved in the content of the tag. This is a var as it
          * may be changed by an xml:space attribute.
          */
-        private var preserveWhitespace = xmlDescriptor.preserveSpace
+        protected var preserveWhitespace = preserveWhitespace
+            private set
 
         protected val attrCount: Int = if (input.eventType == EventType.START_ELEMENT) input.attributeCount else 0
         protected val tagDepth: Int = input.depth
@@ -827,7 +852,7 @@ internal open class XmlDecoderBase internal constructor(
             val isValueChild = index == xmlDescriptor.getValueChild()
             return when (effectiveDeserializer.descriptor.kind) {
                 is PrimitiveKind ->
-                    XmlDecoder(childXmlDescriptor, currentPolyInfo, isValueChild, lastAttrIndex)
+                    XmlDecoder(childXmlDescriptor, currentPolyInfo, isValueChild, lastAttrIndex, preserveWhitespace)
 
                 else -> {
                     SerialValueDecoder(
@@ -836,7 +861,8 @@ internal open class XmlDecoderBase internal constructor(
                         currentPolyInfo,
                         lastAttrIndex,
                         null,
-                        isValueChild
+                        isValueChild,
+                        preserveWhitespace,
                     )
                 }
             }
@@ -870,12 +896,12 @@ internal open class XmlDecoderBase internal constructor(
                 stage == STAGE_NULLS -> NullDecoder(childXmlDescriptor, isValueChild)
 
                 lastAttrIndex >= 0 && childXmlDescriptor is XmlAttributeMapDescriptor -> {
-                    AttributeMapDecoder(effectiveDeserializer, childXmlDescriptor, lastAttrIndex)
+                    AttributeMapDecoder(effectiveDeserializer, childXmlDescriptor, lastAttrIndex, preserveWhitespace)
                 }
 
                 // handle empty value children separately
                 isValueChild && input.hasPeekItems && input.peekNextEvent() == EventType.END_ELEMENT ->
-                    StringDecoder(childXmlDescriptor, input.extLocationInfo, "")
+                    StringDecoder(childXmlDescriptor, input.extLocationInfo, "", preserveWhitespace)
 
                 else -> {
                     serialElementDecoder(descriptor, index, effectiveDeserializer)
@@ -992,7 +1018,7 @@ internal open class XmlDecoderBase internal constructor(
             val childXmlDescriptor = xmlDescriptor.getElementDescriptor(index)
             val isValueChild = index == xmlDescriptor.getValueChild()
             return when (descriptor.kind) {
-                is PrimitiveKind -> XmlDecoder(childXmlDescriptor, currentPolyInfo, isValueChild, lastAttrIndex)
+                is PrimitiveKind -> XmlDecoder(childXmlDescriptor, currentPolyInfo, isValueChild, lastAttrIndex, preserveWhitespace)
 
                 else -> {
                     SerialValueDecoder(
@@ -1001,7 +1027,8 @@ internal open class XmlDecoderBase internal constructor(
                         currentPolyInfo,
                         lastAttrIndex,
                         typeDiscriminatorName,
-                        isValueChild
+                        isValueChild,
+                        preserveWhitespace,
                     )
                 }
             }
@@ -1210,8 +1237,8 @@ internal open class XmlDecoderBase internal constructor(
                         return decodeElementIndex()
                     } else if (attrNS == XML_NS_URI && attrLName == "space") {
                         when (input.getAttributeValue(attrIdx)) {
-                            "preserve" -> preserveWhitespace = true
-                            "default" -> preserveWhitespace = xmlDescriptor.preserveSpace
+                            "preserve" -> preserveWhitespace = DocumentPreserveSpace.DOCUMENT_PRESERVE
+                            "default" -> preserveWhitespace = DocumentPreserveSpace.DEFAULT
                         }
                         // If this was explicitly declared as attribute use that as index, otherwise
                         // just skip the attribute.
@@ -1283,16 +1310,20 @@ internal open class XmlDecoderBase internal constructor(
                         EventType.TEXT -> {
                             // The android reader doesn't check whitespaceness. This code should throw
                             if (input.isWhitespace()) {
-                                if (valueChild != CompositeDecoder.UNKNOWN_NAME && preserveWhitespace) {
+                                if (valueChild != CompositeDecoder.UNKNOWN_NAME) {
                                     var valueDesc = xmlDescriptor.getElementDescriptor(valueChild)
                                     while (valueDesc is XmlListDescriptor && valueDesc.isListEluded) {
                                         valueDesc = valueDesc.getElementDescriptor(0)
                                     }
-                                    if (valueDesc.preserveSpace) { // if the type is not explicitly marked to ignore whitespace
-                                        val outputKind = valueDesc.outputKind
-                                        if (outputKind == OutputKind.Text || outputKind == OutputKind.Mixed) { // this allows all primitives (
-                                            seenItems[valueChild] = true
-                                            return valueChild // We can handle whitespace
+                                    val actualPreserveWS = preserveWhitespace.withDefault(valueDesc.defaultPreserveSpace)
+
+                                    if (actualPreserveWS.withDefault(true)) {
+                                        if (valueDesc.defaultPreserveSpace.withDefault(true)) { // if the type is not explicitly marked to ignore whitespace
+                                            val outputKind = valueDesc.outputKind
+                                            if (outputKind == OutputKind.Text || outputKind == OutputKind.Mixed) { // this allows all primitives (
+                                                seenItems[valueChild] = true
+                                                return valueChild // We can handle whitespace
+                                            }
                                         }
                                     }
                                 }
@@ -1452,16 +1483,18 @@ internal open class XmlDecoderBase internal constructor(
                 OutputKind.Element -> input.readSimpleElement()
 
                 OutputKind.Mixed,
-                OutputKind.Text -> {
-                    input.allConsecutiveTextContent().also { // add some checks that we only have text content
-                        val peek = input.peekNextEvent()
-                        if (peek != EventType.END_ELEMENT) {
-                            throw XmlSerialException("Missing end tag after text only content (found: ${peek})")
-                        } /*else if (peek.localName != serialName.localPart) {
+                OutputKind.Text -> when {
+                    xmlDescriptor.defaultPreserveSpace.withDefault(true) -> input.allConsecutiveTextContent()
+                    else -> input.allText()
+                }.also { // add some checks that we only have text content
+                    val peek = input.peekNextEvent()
+                    if (peek != EventType.END_ELEMENT) {
+                        throw XmlSerialException("Missing end tag after text only content (found: ${peek})")
+                    } /*else if (peek.localName != serialName.localPart) {
                             throw XmlSerialException("Expected end tag local name ${serialName.localPart}, found ${peek.localName}")
                         }*/
-                    }
                 }
+
 
                 OutputKind.Attribute -> error("Attributes should already be read now")
             }
@@ -1476,7 +1509,7 @@ internal open class XmlDecoderBase internal constructor(
             val stringValue = decodeStringElementCollapsed(descriptor, index)
             return when {
                 config.policy.isStrictBoolean -> XmlBooleanSerializer.deserialize(
-                    StringDecoder(xmlDescriptor.getElementDescriptor(index), startPos, stringValue)
+                    StringDecoder(xmlDescriptor.getElementDescriptor(index), startPos, stringValue, preserveWhitespace)
                 )
 
                 else -> stringValue.toBoolean()
@@ -1516,8 +1549,9 @@ internal open class XmlDecoderBase internal constructor(
     internal inner class AttributeMapDecoder(
         deserializer: DeserializationStrategy<*>,
         xmlDescriptor: XmlAttributeMapDescriptor,
-        private val attrIndex: Int
-    ) : TagDecoderBase<XmlAttributeMapDescriptor>(deserializer, xmlDescriptor, null), Decoder {
+        private val attrIndex: Int,
+        inheritedPreserveWhitespace: DocumentPreserveSpace
+    ) : TagDecoderBase<XmlAttributeMapDescriptor>(deserializer, xmlDescriptor, null, inheritedPreserveWhitespace), Decoder {
 
         private var correctStartIndex = -1
         private var nextIndex: Int = 0
@@ -1553,7 +1587,7 @@ internal open class XmlDecoderBase internal constructor(
             val startPos = input.extLocationInfo
             val value = decodeStringElement(descriptor, index)
 
-            val decoder = StringDecoder(xmlDescriptor.valueDescriptor, startPos, value)
+            val decoder = StringDecoder(xmlDescriptor.valueDescriptor, startPos, value, preserveWhitespace)
             return when (effectiveDeserializer) {
                 is XmlDeserializationStrategy ->
                     effectiveDeserializer.deserializeXML(decoder, XmlStringReader(startPos, value))
@@ -1617,8 +1651,9 @@ internal open class XmlDecoderBase internal constructor(
     internal abstract inner class TextualListDecoder(
         deserializer: DeserializationStrategy<*>,
         xmlDescriptor: XmlListDescriptor,
-        private val locationInfo: XmlReader.LocationInfo?
-    ) : TagDecoderBase<XmlListDescriptor>(deserializer, xmlDescriptor, null) {
+        private val locationInfo: XmlReader.LocationInfo?,
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : TagDecoderBase<XmlListDescriptor>(deserializer, xmlDescriptor, null, inheritedPreserveWhitespace) {
         private var listIndex = 0
 
         private val textValues by lazy {
@@ -1641,7 +1676,7 @@ internal open class XmlDecoderBase internal constructor(
             previousValue: T?
         ): T {
             val decoder =
-                StringDecoder(xmlDescriptor.getElementDescriptor(index), locationInfo, textValues[listIndex++])
+                StringDecoder(xmlDescriptor.getElementDescriptor(index), locationInfo, textValues[listIndex++], preserveWhitespace)
             return decoder.decodeSerializableValue(deserializer)
         }
 
@@ -1658,8 +1693,9 @@ internal open class XmlDecoderBase internal constructor(
         deserializer: DeserializationStrategy<*>,
         xmlDescriptor: XmlListDescriptor,
         locationInfo: XmlReader.LocationInfo?,
-        private val attrIndex: Int
-    ) : TextualListDecoder(deserializer, xmlDescriptor, locationInfo) {
+        private val attrIndex: Int,
+        inheritedPreserveWhitespace: DocumentPreserveSpace
+    ) : TextualListDecoder(deserializer, xmlDescriptor, locationInfo, inheritedPreserveWhitespace) {
 
         override fun getTextValue(): String = input.getAttributeValue(attrIndex)
     }
@@ -1667,8 +1703,9 @@ internal open class XmlDecoderBase internal constructor(
     internal inner class ValueListDecoder(
         deserializer: DeserializationStrategy<*>,
         xmlDescriptor: XmlListDescriptor,
-        locationInfo: XmlReader.LocationInfo?
-    ) : TextualListDecoder(deserializer, xmlDescriptor, locationInfo) {
+        locationInfo: XmlReader.LocationInfo?,
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : TextualListDecoder(deserializer, xmlDescriptor, locationInfo, inheritedPreserveWhitespace) {
 
         override fun getTextValue(): String = input.text
     }
@@ -1680,7 +1717,8 @@ internal open class XmlDecoderBase internal constructor(
         private val polyInfo: PolyInfo?,
         typeDiscriminatorName: QName?,
         private val isValueChild: Boolean,
-    ) : TagDecoderBase<XmlListDescriptor>(deserializer, xmlDescriptor, typeDiscriminatorName) {
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : TagDecoderBase<XmlListDescriptor>(deserializer, xmlDescriptor, typeDiscriminatorName, inheritedPreserveWhitespace) {
 
         private val parentXmlDescriptor: XmlDescriptor get() = xmlDescriptor.tagParent.descriptor as XmlDescriptor
 
@@ -1722,7 +1760,8 @@ internal open class XmlDecoderBase internal constructor(
                 polyInfo,
                 Int.MIN_VALUE,
                 typeDiscriminatorName,
-                isValueChild
+                isValueChild,
+                preserveWhitespace,
             )
 
             val result = deserializer.deserializeSafe(decoder, previousValue, false && isValueChild)
@@ -1748,8 +1787,9 @@ internal open class XmlDecoderBase internal constructor(
     internal inner class NamedListDecoder(
         deserializer: DeserializationStrategy<*>,
         xmlDescriptor: XmlListDescriptor,
-        typeDiscriminatorName: QName?
-    ) : TagDecoderBase<XmlListDescriptor>(deserializer, xmlDescriptor, typeDiscriminatorName) {
+        typeDiscriminatorName: QName?,
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : TagDecoderBase<XmlListDescriptor>(deserializer, xmlDescriptor, typeDiscriminatorName, inheritedPreserveWhitespace) {
 
         private var childCount = 0
 
@@ -1777,7 +1817,8 @@ internal open class XmlDecoderBase internal constructor(
                 super.currentPolyInfo,
                 super.lastAttrIndex,
                 null,
-                false
+                false,
+                preserveWhitespace,
             )
 
             // TODO make merging more reliable
@@ -1804,7 +1845,8 @@ internal open class XmlDecoderBase internal constructor(
         xmlDescriptor: XmlMapDescriptor,
         private val polyInfo: PolyInfo?,
         typeDiscriminatorName: QName?,
-    ) : TagDecoderBase<XmlMapDescriptor>(deserializer, xmlDescriptor, typeDiscriminatorName) {
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : TagDecoderBase<XmlMapDescriptor>(deserializer, xmlDescriptor, typeDiscriminatorName, inheritedPreserveWhitespace) {
 
         protected var lastIndex: Int = -1
 
@@ -1825,7 +1867,7 @@ internal open class XmlDecoderBase internal constructor(
                             input.extLocationInfo
                         )
 
-                    val decoder = StringDecoder(keyDescriptor, input.extLocationInfo, key)
+                    val decoder = StringDecoder(keyDescriptor, input.extLocationInfo, key, preserveWhitespace)
 
                     return deserializer.deserializeSafe(decoder, previousValue)
 
@@ -1847,7 +1889,8 @@ internal open class XmlDecoderBase internal constructor(
                 polyInfo,
                 Int.MIN_VALUE,
                 typeDiscriminatorName,
-                false // TODO: this might need to be more specific.
+                false, // TODO: this might need to be more specific.
+                preserveWhitespace,
             )
             if (xmlDescriptor.isValueCollapsed) {
                 decoder.ignoreAttribute(keyDescriptor.tagName)
@@ -1872,7 +1915,8 @@ internal open class XmlDecoderBase internal constructor(
         xmlDescriptor: XmlMapDescriptor,
         polyInfo: PolyInfo?,
         typeDiscriminatorName: QName?,
-    ) : MapDecoderBase(deserializer, xmlDescriptor, polyInfo, typeDiscriminatorName) {
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : MapDecoderBase(deserializer, xmlDescriptor, polyInfo, typeDiscriminatorName, inheritedPreserveWhitespace) {
 
 
         override fun decodeElementIndex(): Int {
@@ -1924,8 +1968,9 @@ internal open class XmlDecoderBase internal constructor(
         deserializer: DeserializationStrategy<*>,
         xmlDescriptor: XmlMapDescriptor,
         polyInfo: PolyInfo?,
-        typeDiscriminatorName: QName?
-    ) : MapDecoderBase(deserializer, xmlDescriptor, polyInfo, typeDiscriminatorName) {
+        typeDiscriminatorName: QName?,
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : MapDecoderBase(deserializer, xmlDescriptor, polyInfo, typeDiscriminatorName, inheritedPreserveWhitespace) {
         override fun Int.checkRepeat(): Int = this
 
         override fun decodeElementIndex(): Int {
@@ -2011,7 +2056,8 @@ internal open class XmlDecoderBase internal constructor(
         xmlDescriptor: XmlPolymorphicDescriptor,
         private val polyInfo: PolyInfo?,
         private val isValueChild: Boolean,
-    ) : TagDecoderBase<XmlPolymorphicDescriptor>(deserializer, xmlDescriptor, null) {
+        inheritedPreserveWhitespace: DocumentPreserveSpace,
+    ) : TagDecoderBase<XmlPolymorphicDescriptor>(deserializer, xmlDescriptor, null, inheritedPreserveWhitespace) {
 
         private var nextIndex = 0
         private var detectedPolyType: String? = null
@@ -2039,7 +2085,8 @@ internal open class XmlDecoderBase internal constructor(
                                 val sdec = StringDecoder(
                                     xmlDescriptor.getElementDescriptor(0),
                                     input.extLocationInfo,
-                                    input.getAttributeValue(i)
+                                    input.getAttributeValue(i),
+                                    preserveWhitespace
                                 )
 
                                 // The QName corresponding to the type serialized
@@ -2097,7 +2144,13 @@ internal open class XmlDecoderBase internal constructor(
                     !xmlDescriptor.isTransparent ->
                         throw XmlSerialException("NonTransparent polymorphic values cannot have text content only")
 
-                    isMixed -> input.allConsecutiveTextContent()
+                    isMixed -> {
+                        if (xmlDescriptor.defaultPreserveSpace.withDefault(true)) {
+                            input.allConsecutiveTextContent()
+                        } else {
+                            input.allText()
+                        }
+                    }
 
                     else -> super.decodeStringElement(descriptor, index)
                 }
@@ -2121,7 +2174,8 @@ internal open class XmlDecoderBase internal constructor(
                 currentPolyInfo,
                 lastAttrIndex,
                 polyTypeAttrname,
-                isValueChild
+                isValueChild,
+                preserveWhitespace,
             )
         }
 
@@ -2142,7 +2196,8 @@ internal open class XmlDecoderBase internal constructor(
                     currentPolyInfo,
                     lastAttrIndex,
                     polyTypeAttrname,
-                    isValueChild
+                    isValueChild,
+                    preserveWhitespace,
                 )
                 nextIndex = 2
                 val result = when (deserializer) {
@@ -2171,7 +2226,7 @@ internal open class XmlDecoderBase internal constructor(
             return when {
                 isMixed && deserializer.descriptor.kind is PrimitiveKind -> {
                     val childXmlDescriptor = xmlDescriptor.getPolymorphicDescriptor(deserializer.descriptor)
-                    deserializer.deserialize(XmlDecoder(childXmlDescriptor, isValueChild = isValueChild))
+                    deserializer.deserialize(XmlDecoder(childXmlDescriptor, isValueChild = isValueChild, inheritedPreserveWhitespace = preserveWhitespace))
                 }
 
                 else ->